Update patch set 2
Patch Set 2:

To create a new version, do all the necessary changes to your code and CI files, submit to Gerrit, wait for the verify job to pass, then merge the change. The build-merge job will push your package to the *staging* repository at PackageCloud.

To release a version, create a releases/*.yaml file and submit it in a new changeset.  Wait for the verify job to pass, then merge the change. The release-merge job will move your package from the staging repository to the *release* repository at PackageCloud.

HTH

Patch-set: 2